consider case where datetime was parsed from the beginning (was never a string)
